feat: add --next feature
When --next is set, the program determines the next major, minor or
patch version.

The possible values for --next are 'major', 'minor' or 'patch'.

- When 'major' is used, the major part of the current semver is bumped
and the minor and patch parts are reset to 0.
- When 'minor' is used, the minor part of the current semver is bumped
and the patch part is reset to 0.
- When 'patch' is used, the patch part of the current semver is bumped.

The next semver only contains the major.minor.patch parts.

Signed-off-by: Thulio Ferraz Assis <thulio.assis@suse.com>